Replaced filter types
