+++ app/bundles/DynamicContentBundle/Assets/js/dynamicContent.js
@@ -24,7 +24,7 @@ Mautic.standardDynamicContentUrl = function(options) {
 };
 
 Mautic.disabledDynamicContentAction = function(opener) {
-    if (typeof opener == 'undefined') {
+    if (typeof opener === 'undefined') {
         opener = window;
     }
 
@@ -35,245 +35,14 @@ Mautic.disabledDynamicContentAction = function(opener) {
     opener.mQuery('#campaignevent_properties_editDynamicContentButton').prop('disabled', disabled);
 };
 
-Mautic.addDwcFilter = function (selectedFilter) {
-    var dwcFilterContainer = mQuery('div.dwc-filter').find('div[data-filter-container]');
-    var selectedOption     = mQuery('option[data-mautic="available_' + selectedFilter + '"]').first();
-    var label              = selectedOption.text();
-
-    // create a new filter
-    var filterNum   = dwcFilterContainer.children('.panel').length;
-    var prototype   = mQuery('#filterSelectPrototype').data('prototype');
-    var fieldObject = selectedOption.data('field-object');
-    var fieldType   = selectedOption.data('field-type');
-    var isSpecial   = (mQuery.inArray(fieldType, ['leadlist', 'lead_email_received', 'tags', 'multiselect', 'boolean', 'select', 'country', 'timezone', 'region', 'stage', 'locale']) != -1);
-
-    // Update the prototype settings
-    prototype = prototype
-        .replace(/__name__/g, filterNum)
-        .replace(/__label__/g, label)
-        .replace(/filters_0_filters/g, 'filters')
-        .replace(/filters]\[0]\[filters/g, 'filters');
-
-    if (filterNum === 0) {
-        prototype = prototype.replace(/in-group/g, '');
-    }
-
-    // Convert to DOM
-    prototype = mQuery(prototype);
-
-    if (fieldObject === 'company') {
-        prototype.find('.object-icon').removeClass('fa-user').addClass('fa-building');
-    } else {
-        prototype.find('.object-icon').removeClass('fa-building').addClass('fa-user');
-    }
-
-    var filterBase  = "dwc[filters][" + filterNum + "]";
-    var filterIdBase = "dwc_filters_" + filterNum;
-
-    if (isSpecial) {
-        var templateField = fieldType;
-        if (fieldType === 'boolean' || fieldType === 'multiselect') {
-            templateField = 'select';
-        }
-        var template = mQuery('#templates .' + templateField + '-template').clone();
-        var $template = mQuery(template);
-        var templateNameAttr = $template.attr('name').replace(/__name__/g, filterNum);
-        var templateIdAttr = $template.attr('id').replace(/__name__/g, filterNum);
-
-        $template.attr('name', templateNameAttr);
-        $template.attr('id', templateIdAttr);
-
-        prototype.find('input[name="' + filterBase + '[filter]"]').replaceWith(template);
-    }
-
-    if (dwcFilterContainer.find('.panel').length === 0) {
-        // First filter so hide the glue footer
-        prototype.find(".panel-footer").addClass('hide');
-    }
-
-    prototype.find("input[name='" + filterBase + "[field]']").val(selectedFilter);
-    prototype.find("input[name='" + filterBase + "[type]']").val(fieldType);
-    prototype.find("input[name='" + filterBase + "[object]']").val(fieldObject);
-
-    var filterEl = (isSpecial) ? "select[name='" + filterBase + "[filter]']" : "input[name='" + filterBase + "[filter]']";
-
-    dwcFilterContainer.append(prototype);
-
-    Mautic.initRemoveEvents(dwcFilterContainer.find("a.remove-selected"));
-
-    var filter = '#' + filterIdBase + '_filter';
-
-    var fieldOptions = fieldCallback = '';
-    //activate fields
-    if (isSpecial) {
-        if (fieldType === 'select' || fieldType === 'boolean' || fieldType === 'multiselect') {
-            // Generate the options
-            fieldOptions = selectedOption.data("field-list");
-
-            mQuery.each(fieldOptions, function(index, val) {
-                mQuery('<option>').val(index).text(val).appendTo(filterEl);
-            });
-        }
-    } else if (fieldType === 'lookup') {
-        fieldCallback = selectedOption.data("field-callback");
-        if (fieldCallback && typeof Mautic[fieldCallback] === 'function') {
-            fieldOptions = selectedOption.data("field-list");
-            Mautic[fieldCallback](filterIdBase + '_filter', selectedFilter, fieldOptions);
-        }
-    } else if (fieldType === 'datetime') {
-        mQuery(filter).datetimepicker({
-            format: 'Y-m-d H:i',
-            lazyInit: true,
-            validateOnBlur: false,
-            allowBlank: true,
-            scrollInput: false
-        });
-    } else if (fieldType === 'date') {
-        mQuery(filter).datetimepicker({
-            timepicker: false,
-            format: 'Y-m-d',
-            lazyInit: true,
-            validateOnBlur: false,
-            allowBlank: true,
-            scrollInput: false,
-            closeOnDateSelect: true
-        });
-    } else if (fieldType === 'time') {
-        mQuery(filter).datetimepicker({
-            datepicker: false,
-            format: 'H:i',
-            lazyInit: true,
-            validateOnBlur: false,
-            allowBlank: true,
-            scrollInput: false
-        });
-    } else if (fieldType === 'lookup_id') {
-        //switch the filter and display elements
-        var oldFilter = mQuery(filterEl);
-        var newDisplay = mQuery(oldFilter).clone();
-        mQuery(newDisplay).attr('name', filterBase + '[display]')
-            .attr('id', filterIdBase + '_display');
-
-        var oldDisplay = mQuery(prototype).find("input[name='" + filterBase + "[display]']");
-        var newFilter = mQuery(oldDisplay).clone();
-        mQuery(newFilter).attr('name', filterBase + '[filter]')
-            .attr('id', filterIdBase + '_filter');
-
-        mQuery(oldFilter).replaceWith(newFilter);
-        mQuery(oldDisplay).replaceWith(newDisplay);
-
-        var fieldCallback = selectedOption.data("field-callback");
-        if (fieldCallback && typeof Mautic[fieldCallback] === 'function') {
-            fieldOptions = selectedOption.data("field-list");
-            Mautic[fieldCallback](filterIdBase + '_display', selectedFilter, fieldOptions);
-        }
-    } else {
-        mQuery(filter).attr('type', fieldType);
-    }
-
-    var operators = mQuery(selectedOption).data('field-operators');
-    mQuery('#' + filterIdBase + '_operator').html('');
-    mQuery.each(operators, function (value, label) {
-        var newOption = mQuery('<option/>').val(value).text(label);
-        newOption.appendTo(mQuery('#' + filterIdBase + '_operator'));
-    });
-
-    // Convert based on first option in list
-    Mautic.convertDwcFilterInput('#' + filterIdBase + '_operator');
-};
-
-Mautic.convertDwcFilterInput = function(el) {
-    var operator = mQuery(el).val();
-    // Extract the filter number
-    var regExp    = /dwc_filters_(\d+)_operator/;
-    var matches   = regExp.exec(mQuery(el).attr('id'));
-
-    var filterNum                 = matches[1];
-
-    var filterId       = '#dwc_filters_' + filterNum + '_filter';
-    var filterEl       = mQuery(filterId);
-    var filterElParent = filterEl.parent();
-
-    // Reset has-error
-    if (filterElParent.hasClass('has-error')) {
-        filterElParent.find('div.help-block').hide();
-        filterElParent.removeClass('has-error');
-    }
-
-    var disabled = (operator === 'empty' || operator === '!empty');
-    filterEl.prop('disabled', disabled);
-
-    if (disabled) {
-        filterEl.val('');
-    }
-
-    var newName = '';
-    var lastPos;
-
-    if (filterEl.is('select')) {
-        var isMultiple  = filterEl.attr('multiple');
-        var multiple    = (operator === 'in' || operator === '!in');
-        var placeholder = filterEl.attr('data-placeholder');
-
-        if (multiple && !isMultiple) {
-            filterEl.attr('multiple', 'multiple');
-
-            // Update the name
-            newName =  filterEl.attr('name') + '[]';
-            filterEl.attr('name', newName);
-
-            placeholder = mauticLang['chosenChooseMore'];
-        } else if (!multiple && isMultiple) {
-            filterEl.removeAttr('multiple');
-
-            // Update the name
-            newName = filterEl.attr('name');
-            lastPos = newName.lastIndexOf('[]');
-            newName = newName.substring(0, lastPos);
-
-            filterEl.attr('name', newName);
-
-            placeholder = mauticLang['chosenChooseOne'];
-        }
-
-        if (multiple) {
-            // Remove empty option
-            filterEl.find('option[value=""]').remove();
-
-            // Make sure none are selected
-            filterEl.find('option:selected').removeAttr('selected');
-        } else {
-            // Add empty option
-            filterEl.prepend("<option value='' selected></option>");
-        }
-
-        // Destroy the chosen and recreate
-        if (mQuery(filterId + '_chosen').length) {
-            filterEl.chosen('destroy');
-        }
-
-        filterEl.attr('data-placeholder', placeholder);
-
-        Mautic.activateChosenSelect(filterEl, false);
-    }
-};
-
-Mautic.onDwcChange = function(el) {
-    var $this = mQuery(el);
-
-    if ($this.val()) {
-        Mautic.addDwcFilter($this.val());
-        $this.val('');
-        $this.trigger('chosen:updated');
-    }
-};
-
 if (typeof MauticIsDwcReady === 'undefined') {
     var MauticIsDwcReady = true;
     // Handler when the DOM is fully loaded
     var callback = function(){
         var availableFilters = mQuery('div.dwc-filter').find('select[data-mautic="available_filters"]');
         Mautic.activateChosenSelect(availableFilters, false);
+
+        Mautic.leadlistOnLoad('div.dwc-filter');
     };
 
     if (
