+++ app/bundles/DynamicContentBundle/Form/Type/DynamicContentType.php
@@ -11,33 +11,132 @@
 
 namespace Mautic\DynamicContentBundle\Form\Type;
 
+use DeviceDetector\Parser\Device\DeviceParserAbstract as DeviceParser;
+use DeviceDetector\Parser\OperatingSystem;
 use Doctrine\ORM\EntityManager;
+use Mautic\CategoryBundle\Form\Type\CategoryListType;
+use Mautic\CategoryBundle\Model\CategoryModel;
 use Mautic\CoreBundle\Form\DataTransformer\EmojiToShortTransformer;
 use Mautic\CoreBundle\Form\DataTransformer\IdToEntityModelTransformer;
 use Mautic\CoreBundle\Form\EventListener\CleanFormSubscriber;
 use Mautic\CoreBundle\Form\EventListener\FormExitSubscriber;
-use Mautic\CoreBundle\Form\Type\DwcFiltersTrait;
+use Mautic\CoreBundle\Form\Type\FormButtonsType;
+use Mautic\CoreBundle\Form\Type\YesNoButtonGroupType;
+use Mautic\CoreBundle\Helper\UserHelper;
+use Mautic\CoreBundle\Security\Permissions\CorePermissions;
+use Mautic\DynamicContentBundle\Entity\DynamicContent;
+use Mautic\EmailBundle\Model\EmailModel;
+use Mautic\LeadBundle\Form\DataTransformer\FieldFilterTransformer;
+use Mautic\LeadBundle\Helper\FormFieldHelper;
+use Mautic\LeadBundle\Model\LeadModel;
+use Mautic\LeadBundle\Model\ListModel;
+use Mautic\StageBundle\Model\StageModel;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CollectionType;
+use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
+use Symfony\Component\Form\Extension\Core\Type\LocaleType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Translation\TranslatorInterface;
 
 /**
  * Class DynamicContentType.
  */
 class DynamicContentType extends AbstractType
 {
-    use DwcFiltersTrait;
-
     private $em;
+
+    private $fieldChoices;
+    private $countryChoices;
+    private $regionChoices;
+    private $timezoneChoices;
+    private $stageChoices = [];
+    private $localeChoices;
+
+    private $translator;
+    private $listChoices  = [];
+    private $emailChoices = [];
+    private $deviceTypesChoices;
+    private $deviceBrandsChoices;
+    private $deviceOsChoices;
+    private $tagChoices        = [];
+    private $categoriesChoices = [];
 
     /**
      * DynamicContentType constructor.
      *
-     * @param EntityManager $entityManager
+     * @param EntityManager       $entityManager
+     * @param ListModel           $listModel
+     * @param StageModel          $stageModel
+     * @param TranslatorInterface $translator
+     * @param LeadModel           $leadModel
+     * @param CategoryModel       $categoryModel
+     * @param EmailModel          $emailModel
+     * @param CorePermissions     $security
+     * @param UserHelper          $userHelper
+     *
+     * @throws \InvalidArgumentException
      */
-    public function __construct(EntityManager $entityManager)
+    public function __construct(EntityManager $entityManager, ListModel $listModel, StageModel $stageModel, TranslatorInterface $translator, LeadModel $leadModel, CategoryModel $categoryModel, EmailModel $emailModel, CorePermissions $security, UserHelper $userHelper)
     {
         $this->em = $entityManager;
+
+        $this->fieldChoices = $listModel->getChoiceFields();
+
+        $this->countryChoices  = FormFieldHelper::getCountryChoices();
+        $this->regionChoices   = FormFieldHelper::getRegionChoices();
+        $this->timezoneChoices = FormFieldHelper::getTimezonesChoices();
+        $this->localeChoices   = FormFieldHelper::getLocaleChoices();
+
+        $stages = $stageModel->getRepository()->getSimpleList();
+        foreach ($stages as $stage) {
+            $this->stageChoices[$stage['value']] = $stage['label'];
+        }
+
+        $this->translator = $translator;
+
+        // emails
+        $viewOther   = $security->isGranted('email:emails:viewother');
+        $currentUser = $userHelper->getUser();
+        $emailRepo   = $emailModel->getRepository();
+
+        $emailRepo->setCurrentUser($currentUser);
+
+        $emails = $emailRepo->getEmailList('', 0, 0, $viewOther, true);
+
+        foreach ($emails as $email) {
+            $this->emailChoices[$email['language']][$email['id']] = $email['name'];
+        }
+        ksort($this->emailChoices);
+
+        $tags = $leadModel->getTagList();
+        foreach ($tags as $tag) {
+            $this->tagChoices[$tag['value']] = $tag['label'];
+        }
+
+        $categories = $categoryModel->getLookupResults('global');
+        foreach ($categories as $category) {
+            $this->categoriesChoices[$category['id']] = $category['title'];
+        }
+
+        // Segments
+        /** @var array $lists */
+        $lists = $listModel->getUserLists();
+        foreach ($lists as $list) {
+            $this->listChoices[$list['id']] = $list['name'];
+        }
+
+        $this->deviceTypesChoices  = array_combine(DeviceParser::getAvailableDeviceTypeNames(), DeviceParser::getAvailableDeviceTypeNames());
+        $this->deviceBrandsChoices = DeviceParser::$deviceBrands;
+        $this->deviceOsChoices     = array_combine(
+            array_keys(OperatingSystem::getAvailableOperatingSystemFamilies()),
+            array_keys(OperatingSystem::getAvailableOperatingSystemFamilies())
+        );
     }
 
     /**
@@ -51,7 +150,7 @@ class DynamicContentType extends AbstractType
 
         $builder->add(
             'name',
-            'text',
+            TextType::class,
             [
                 'label'      => 'mautic.dynamicContent.form.internal.name',
                 'label_attr' => ['class' => 'control-label'],
@@ -63,7 +162,7 @@ class DynamicContentType extends AbstractType
         $builder->add(
             $builder->create(
                 'description',
-                'textarea',
+                TextareaType::class,
                 [
                     'label'      => 'mautic.dynamicContent.description',
                     'label_attr' => ['class' => 'control-label'],
@@ -73,11 +172,11 @@ class DynamicContentType extends AbstractType
             )->addModelTransformer($emojiTransformer)
         );
 
-        $builder->add('isPublished', 'yesno_button_group');
+        $builder->add('isPublished', YesNoButtonGroupType::class);
 
         $builder->add(
             'language',
-            'locale',
+            LocaleType::class,
             [
                 'label'      => 'mautic.core.language',
                 'label_attr' => ['class' => 'control-label'],
@@ -90,7 +189,7 @@ class DynamicContentType extends AbstractType
 
         $builder->add(
             'publishUp',
-            'datetime',
+            DateTimeType::class,
             [
                 'widget'     => 'single_text',
                 'label'      => 'mautic.core.form.publishup',
@@ -106,7 +205,7 @@ class DynamicContentType extends AbstractType
 
         $builder->add(
             'publishDown',
-            'datetime',
+            DateTimeType::class,
             [
                 'widget'     => 'single_text',
                 'label'      => 'mautic.core.form.publishdown',
@@ -122,7 +221,7 @@ class DynamicContentType extends AbstractType
 
         $builder->add(
             'content',
-            'textarea',
+            TextareaType::class,
             [
                 'label'      => 'mautic.dynamicContent.form.content',
                 'label_attr' => ['class' => 'control-label'],
@@ -139,7 +238,7 @@ class DynamicContentType extends AbstractType
         $builder->add(
             $builder->create(
                 'translationParent',
-                'dwc_list',
+                DynamicContentListType::class,
                 [
                     'label'      => 'mautic.core.form.translation_parent',
                     'label_attr' => ['class' => 'control-label'],
@@ -158,20 +257,20 @@ class DynamicContentType extends AbstractType
 
         $builder->add(
             'category',
-            'category',
+            CategoryListType::class,
             ['bundle' => 'dynamicContent']
         );
 
         if (!empty($options['update_select'])) {
             $builder->add(
                 'buttons',
-                'form_buttons',
+                FormButtonsType::class,
                 ['apply_text' => false]
             );
 
             $builder->add(
                 'updateSelect',
-                'hidden',
+                HiddenType::class,
                 [
                     'data'   => $options['update_select'],
                     'mapped' => false,
@@ -180,11 +279,39 @@ class DynamicContentType extends AbstractType
         } else {
             $builder->add(
                 'buttons',
-                'form_buttons'
+                FormButtonsType::class
             );
         }
 
-        $this->addFiltersField($builder);
+        $filterModalTransformer = new FieldFilterTransformer($this->translator);
+        $builder->add(
+            $builder->create(
+                'filters',
+                CollectionType::class,
+                [
+                    'type'    => DwcEntryFiltersType::class,
+                    'options' => [
+                        'countries'      => $this->countryChoices,
+                        'regions'        => $this->regionChoices,
+                        'timezones'      => $this->timezoneChoices,
+                        'stages'         => $this->stageChoices,
+                        'locales'        => $this->localeChoices,
+                        'fields'         => $this->fieldChoices,
+                        'emails'         => $this->emailChoices,
+                        'lists'          => $this->listChoices,
+                        'deviceTypes'    => $this->deviceTypesChoices,
+                        'deviceBrands'   => $this->deviceBrandsChoices,
+                        'deviceOs'       => $this->deviceOsChoices,
+                        'tags'           => $this->tagChoices,
+                        'globalcategory' => $this->categoriesChoices,
+                    ],
+                    'error_bubbling' => false,
+                    'mapped'         => true,
+                    'allow_add'      => true,
+                    'allow_delete'   => true,
+                ]
+            )->addModelTransformer($filterModalTransformer)
+        );
 
         if (!empty($options['action'])) {
             $builder->setAction($options['action']);
@@ -193,10 +320,16 @@ class DynamicContentType extends AbstractType
 
     /**
      * @param OptionsResolver $resolver
+     *
+     * @throws \Symfony\Component\OptionsResolver\Exception\AccessException
      */
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults(['data_class' => 'Mautic\DynamicContentBundle\Entity\DynamicContent']);
+        $resolver->setDefaults([
+            'data_class'     => DynamicContent::class,
+            'label'          => false,
+            'error_bubbling' => false,
+        ]);
 
         $resolver->setDefined(['update_select']);
     }
@@ -204,8 +337,28 @@ class DynamicContentType extends AbstractType
     /**
      * @return string
      */
-    public function getName()
+    public function getName(): string
     {
         return 'dwc';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function buildView(FormView $view, FormInterface $form, array $options)
+    {
+        $view->vars['fields']         = $this->fieldChoices;
+        $view->vars['countries']      = $this->countryChoices;
+        $view->vars['regions']        = $this->regionChoices;
+        $view->vars['timezones']      = $this->timezoneChoices;
+        $view->vars['emails']         = $this->emailChoices;
+        $view->vars['lists']          = $this->listChoices;
+        $view->vars['deviceTypes']    = $this->deviceTypesChoices;
+        $view->vars['deviceBrands']   = $this->deviceBrandsChoices;
+        $view->vars['deviceOs']       = $this->deviceOsChoices;
+        $view->vars['tags']           = $this->tagChoices;
+        $view->vars['stage']          = $this->stageChoices;
+        $view->vars['locales']        = $this->localeChoices;
+        $view->vars['globalcategory'] = $this->categoriesChoices;
     }
 }
