+++ app/bundles/DynamicContentBundle/Model/DynamicContentModel.php
@@ -27,6 +27,9 @@ use Mautic\LeadBundle\Entity\Lead;
 use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
+/**
+ * Class DynamicContentModel.
+ */
 class DynamicContentModel extends FormModel implements AjaxLookupModelInterface
 {
     use VariantModelTrait;
